# switch encoding/json to json iter