Signed-off-by: Ben Ye <benye@amazon.com>

<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with master
-->

**What this PR does**:
`SampleStream` will be marshal/unmarshalled in query frontend for instant query response. Right now, it uses `encoding/json` library for json serialization. This PR switches it to json iter library which should provide better performance.

Existing tests should still pass.

Benchmark

UnMarshal

```
name                              old time/op    new time/op    delta
SampleStreamUnmarshalJSONITER-10    1.64ms ± 0%    1.05ms ± 0%   ~     (p=1.000 n=1+1)                          

name                              old alloc/op   new alloc/op   delta
SampleStreamUnmarshalJSONITER-10     934kB ± 0%    1090kB ± 0%   ~     (p=1.000 n=1+1)

name                              old allocs/op  new allocs/op  delta
SampleStreamUnmarshalJSONITER-10      49.0 ± 0%      53.0 ± 0%   ~     (p=1.000 n=1+1)
```

Marshal

```
name                            old time/op    new time/op    delta
SampleStreamMarshalJSONITER-10    2.29ms ± 0%    1.40ms ± 0%   ~     (p=1.000 n=1+1)

name                            old alloc/op   new alloc/op   delta
SampleStreamMarshalJSONITER-10     636kB ± 0%     436kB ± 0%   ~     (p=1.000 n=1+1)

name                            old allocs/op  new allocs/op  delta
SampleStreamMarshalJSONITER-10     20.0k ± 0%     20.0k ± 0%   ~     (p=1.000 n=1+1)
```


**Which issue(s) this PR fixes**:
Fixes #<issue number>

**Checklist**
- [x] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`